+++ app/src/main/java/woowacourse/shopping/ShoppingApplication.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping
 
 import android.app.Application
-import woowacourse.shopping.data.db.cart.CartDatabase
-import woowacourse.shopping.data.db.recentproduct.RecentProductDatabase
+import woowacourse.shopping.data.local.db.cart.CartDatabase
+import woowacourse.shopping.data.local.db.recentproduct.RecentProductDatabase
 
 class ShoppingApplication : Application() {
     override fun onCreate() {
