+++ app/src/main/java/woowacourse/shopping/data/datasource/impl/CouponRemoteDataSourceImpl.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.data.datasource.impl
 
 import woowacourse.shopping.data.datasource.CouponRemoteDataSource
 import woowacourse.shopping.data.dto.response.ResponseCouponDto
-import woowacourse.shopping.data.service.CouponService
+import woowacourse.shopping.data.remote.service.CouponService
 import woowacourse.shopping.exception.ShoppingError
 import woowacourse.shopping.exception.ShoppingException
