+++ app/src/main/java/woowacourse/shopping/data/datasource/impl/OrderRemoteDataSourceImpl.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.data.datasource.impl
 
 import woowacourse.shopping.data.datasource.OrderRemoteDataSource
 import woowacourse.shopping.data.dto.request.RequestOrdersPostDto
-import woowacourse.shopping.data.service.OrderService
+import woowacourse.shopping.data.remote.service.OrderService
 
 class OrderRemoteDataSourceImpl(private val service: OrderService) : OrderRemoteDataSource {
     override suspend fun order(request: RequestOrdersPostDto): Result<Unit> =
