+++ app/src/main/java/woowacourse/shopping/data/local/db/cart/CartDatabase.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.db.cart
+package woowacourse.shopping.data.local.db.cart
 
 import android.content.Context
 import androidx.room.Database
