+++ app/src/main/java/woowacourse/shopping/data/local/db/recentproduct/RecentProductDao.kt
@@ -1,10 +1,10 @@
-package woowacourse.shopping.data.db.recentproduct
+package woowacourse.shopping.data.local.db.recentproduct
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import woowacourse.shopping.data.model.local.RecentProduct
+import woowacourse.shopping.data.model.RecentProduct
 
 @Dao
 interface RecentProductDao {
