+++ app/src/main/java/woowacourse/shopping/data/remote/api/ProductMockWebServerDispatcher.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.api
+package woowacourse.shopping.data.remote.api
 
 import com.google.gson.Gson
 import okhttp3.mockwebserver.Dispatcher
