+++ app/src/main/java/woowacourse/shopping/data/remote/api/ProductServerApi.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.api
+package woowacourse.shopping.data.remote.api
 
 import woowacourse.shopping.domain.model.product.Product
