+++ app/src/main/java/woowacourse/shopping/data/remote/service/CouponService.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.service
+package woowacourse.shopping.data.remote.service
 
 import retrofit2.Response
 import retrofit2.http.GET
