+++ app/src/main/java/woowacourse/shopping/data/remote/service/NetworkModule.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.service
+package woowacourse.shopping.data.remote.service
 
 import okhttp3.OkHttpClient
 import retrofit2.Retrofit
