+++ app/src/main/java/woowacourse/shopping/data/repository/CouponRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.data.repository
 
 import woowacourse.shopping.data.datasource.CouponRemoteDataSource
-import woowacourse.shopping.data.utils.mapper.toDomain
+import woowacourse.shopping.data.utils.toDomain
 import woowacourse.shopping.domain.model.coupon.CouponState
 import woowacourse.shopping.domain.repository.CouponRepository
