+++ app/src/main/java/woowacourse/shopping/ui/cart/CartActivity.kt
@@ -12,11 +12,11 @@ import woowacourse.shopping.data.datasource.impl.CartRemoteDataSourceImpl
 import woowacourse.shopping.data.datasource.impl.OrderRemoteDataSourceImpl
 import woowacourse.shopping.data.datasource.impl.ProductRemoteDataSourceImpl
 import woowacourse.shopping.data.datasource.impl.RecentProductLocalDataSourceImpl
-import woowacourse.shopping.data.db.recentproduct.RecentProductDatabase
+import woowacourse.shopping.data.local.db.recentproduct.RecentProductDatabase
+import woowacourse.shopping.data.remote.service.NetworkModule
 import woowacourse.shopping.data.repository.CartRepositoryImpl
 import woowacourse.shopping.data.repository.ProductRepositoryImpl
 import woowacourse.shopping.data.repository.RecentProductRepositoryImpl
-import woowacourse.shopping.data.service.NetworkModule
 import woowacourse.shopping.databinding.ActivityCartBinding
 import woowacourse.shopping.exception.handleError
 import woowacourse.shopping.ui.cart.adapter.CartAdapter
