+++ app/src/test/java/woowacourse/shopping/ui/FakeRecentProductDao.kt
@@ -1,7 +1,7 @@
 package woowacourse.shopping.ui
 
-import woowacourse.shopping.data.db.recentproduct.RecentProductDao
-import woowacourse.shopping.data.model.local.RecentProduct
+import woowacourse.shopping.data.local.db.recentproduct.RecentProductDao
+import woowacourse.shopping.data.model.RecentProduct
 import java.time.LocalDateTime
 
 object FakeRecentProductDao : RecentProductDao {
